Add changes in libgeometry

+++ src/libgeometry/perimetr.h
@@ -1,3 +1,3 @@
 #pragma once
 
-float calculatePerimetr(char*);
+double calculatePerimetr(char*);
